CLI-561 Move SCA scanning to pre-commit hook
